refactor: move helper functions to services
