+++ backend/src/routers/csv_to_json.py
@@ -1,7 +1,8 @@
 from fastapi import APIRouter, File, UploadFile
 from fastapi.responses import JSONResponse
-import pandas as pd
 from io import StringIO
+
+from src.services import csv_to_json_service as service
 
 router = APIRouter()
 
@@ -12,20 +13,7 @@ async def upload_csv(file: UploadFile = File(...)):
         # Read the CSV file content
         contents = await file.read()
         csv_data = StringIO(contents.decode("utf-8"))
-
-        # Use pandas to read the CSV data
-        df = pd.read_csv(csv_data)
-
-        # Convert DataFrame to JSON
-        df_array = df.to_dict(orient="tight")
-
-        keys = df_array["columns"]
-
-        # Convert DataFrame object to the list of dictionares
-        data_array = [dict(zip(keys, item)) for item in df_array["data"]]
-
-        json_object = {"data": data_array}
-
+        json_object = await service.resolve_json_object(csv_data)
         return JSONResponse(content=json_object)
 
     except Exception as e:
