+++ backend/src/routers/gene_expression.py
@@ -1,6 +1,8 @@
 from fastapi import APIRouter
 import pandas as pd
+
 from src.models import GeneExpressionData
+from src.services import gene_expression_service as service
 
 router = APIRouter()
 
@@ -9,10 +11,7 @@ router = APIRouter()
 async def process_data(gene_expression_data: GeneExpressionData):
     # Convert the incoming data into a DataFrame
     df = pd.DataFrame(gene_expression_data.data)
-
-    # Calculate gene counts
-    gene_count = df["gene"].value_counts()
-    gene_count_dict = gene_count.to_dict()
+    gene_count_dict = await service.get_gene_count(df)
 
     # Prepare the result dictionary
     result = df.describe().to_dict()
