+++ backend/src/routers/sample_csv.py
@@ -1,31 +1,15 @@
 from fastapi import APIRouter, Response
-import csv
-import io
-import random
-from src.utils.gene_names import GENE_NAMES
+
+from src.services import sample_csv_service as service
 
 router = APIRouter()
-
-
-def generate_random_gene_expression(num_records):
-    data = []
-    for _ in range(num_records):
-        gene_id = random.choice(GENE_NAMES)
-        expression_value = round(random.uniform(0.0, 100.0), 2)
-        data.append({"gene": gene_id, "expression_level": expression_value})
-    return data
 
 
 @router.get("/sample-csv")
 async def get_sample_csv():
     num_records = 1000
-    sample_data = generate_random_gene_expression(num_records)
-
-    output = io.StringIO()
-    writer = csv.DictWriter(output, fieldnames=["gene", "expression_level"])
-    writer.writeheader()
-    writer.writerows(sample_data)
-    content = output.getvalue()
+    sample_data = await service.generate_random_gene_expression(num_records)
+    content = await service.resolve_csv_content(sample_data)
 
     return Response(
         content,
