+++ backend/src/services/gene_expression_service.py
@@ -1,0 +1,7 @@
+from pandas import DataFrame
+
+
+async def get_gene_count(df: DataFrame):
+    gene_count = df["gene"].value_counts()
+    gene_count_dict = gene_count.to_dict()
+    return gene_count_dict
